AMB-846 move examples to the specification
