+++ specification/examples/stu3.json
