+++ tests/test_converter.py
@@ -1,7 +1,7 @@
 import pytest
 import requests
 
-from .examples.example_loader import load_example
+from .example_loader import load_example
 
 
 class TestConverter:
@@ -43,13 +43,14 @@ class TestConverter:
         expected_response = load_example("r4_to_stu3_200.json")
         assert res.json() == expected_response
 
+    @pytest.mark.debug
     def test_converter_json_r4_to_xml_r4(self, url, token):
         # Given
-        r4_payload = load_example("r4.xml")
+        r4_payload = load_example("r4.json")
         headers = {
             "Authorization": f"Bearer {token}",
-            "Content-Type": "application/fhir+xml; fhirVersion=4.0",
-            "Accept": "application/fhir+json; fhirVersion=3.0"
+            "Content-Type": "application/fhir+json; fhirVersion=4.0",
+            "Accept": "application/fhir+xml; fhirVersion=4.0"
         }
 
         # When
@@ -57,7 +58,9 @@ class TestConverter:
 
         # Then
         # Fixme: backend returns 500
+        print("foooo")
         print(res.text)
+        print(res.status_code)
         # assert res.status_code == 200
         # expected_response = load_example("r4_to_r4_200.json")
         # assert res.json() == expected_response
